Deduplicate per-version PrintSuite and JavacpSuite test copies
